+++ spec/type_aliases/apics_executionmode_spec.rb
@@ -1,6 +1,0 @@
-require 'spec_helper'
-
-describe 'Apics::ExecutionMode' do
-  it { is_expected.to allow_values('Development', 'Production') }
-  it { is_expected.not_to allow_value('Test') }
-end